fix invoicable patien report page
